refactor(agent): split mcp client responsibilities
