+++ packages/agent/src/runtime/mcp/client-descriptor.ts
@@ -1,0 +1,34 @@
+import type { RuntimeResource, Tool } from "@openomni/protocol";
+
+type McpToolSpec = Tool.Spec & {
+  readonly descriptor: RuntimeResource.Descriptor;
+};
+
+function createMcpToolDescriptor(serverId: string, remoteName: string): RuntimeResource.Descriptor {
+  return {
+    id: `tool:mcp:${serverId}:${remoteName}`,
+    kind: "tool",
+    source: {
+      type: "mcp",
+      serverId,
+      remoteName,
+    },
+    labels: ["source.mcp", `mcp.${serverId}`],
+    capabilities: ["network.write"],
+    effects: ["external.write"],
+  };
+}
+
+export function attachMcpToolDescriptor(
+  spec: Tool.Spec,
+  serverId: string,
+  remoteName: string,
+): McpToolSpec {
+  const descriptor = createMcpToolDescriptor(serverId, remoteName);
+
+  return {
+    ...spec,
+    labels: descriptor.labels,
+    descriptor,
+  };
+}
